Revert "SALTO-1078 convert lists to maps (#2613)"
This reverts commit a65a531476ca2a553ee741ecd2443db52868111b.
