+++ packages/netsuite-adapter/src/adapter.ts
@@ -29,7 +29,6 @@ import {
 import { getMetadataTypes, getTopLevelCustomTypes, metadataTypesToList } from './types'
 import { TYPES_TO_SKIP, FILE_PATHS_REGEX_SKIP_LIST,
   INTEGRATION, FETCH_TARGET, SKIP_LIST, LAST_FETCH_TIME, USE_CHANGES_DETECTION, FETCH, INCLUDE, EXCLUDE, DEPLOY, DEPLOY_REFERENCED_ELEMENTS, WARN_STALE_DATA, APPLICATION_ID, LOCKED_ELEMENTS_TO_EXCLUDE } from './constants'
-import convertListsToMaps from './filters/convert_lists_to_maps'
 import replaceInstanceReferencesFilter from './filters/instance_references'
 import parseSavedSearch from './filters/parse_saved_searchs'
 import convertLists from './filters/convert_lists'
@@ -121,8 +120,6 @@ export default class NetsuiteAdapter implements AdapterOperations {
       addParentFolder,
       parseSavedSearch,
       convertLists,
-      // convertListsToMaps must run after convertLists and before replaceInstanceReferencesFilter
-      convertListsToMaps,
       consistentValues,
       replaceInstanceReferencesFilter,
       SDFInternalIds,
