+++ packages/netsuite-adapter/src/constants.ts
@@ -57,8 +57,6 @@ export const APPLICATION_ID = 'application_id'
 // Field Annotations
 export const IS_ATTRIBUTE = 'isAttribute'
 export const ADDITIONAL_FILE_SUFFIX = 'additionalFileSuffix'
-export const LIST_MAPPED_BY_FIELD = 'map_key_field'
-export const INDEX = 'index'
 
 // SDF FileCabinet top level folders
 export const FILE_CABINET_PATH_SEPARATOR = '/'
